patch

+++ demo/src/main/java/com/example/demo/Aperitivo.java
@@ -1,7 +1,7 @@
 package com.example.demo;
 
 public class Aperitivo extends BebidaAlcoholica{
-    public Aperitivo(String descripcion, String nombre, String marca, String tipo, int cantidad, int precio, boolean alcohol, boolean artesanal, int graduacion) {
+    public Aperitivo(String descripcion, String nombre, String marca, String tipo, int cantidad, int precio, boolean alcohol, boolean artesanal, double graduacion) {
         super(descripcion, nombre, marca, tipo, cantidad, precio, alcohol, artesanal, graduacion);
     }
