+++ demo/src/main/java/com/example/demo/Cerveza.java
@@ -2,16 +2,16 @@ package com.example.demo;
 
 public class Cerveza extends BebidaAlcoholica{
 
-    private int ibu;
+    private double ibu;
     private String color;
 
-    public Cerveza(String color,int ibu,String descripcion, String nombre, String marca, String tipo, int cantidad, int precio, boolean alcohol, boolean artesanal, int graduacion){
+    public Cerveza(String color, double ibu, String descripcion, String nombre, String marca, String tipo, int cantidad, int precio, boolean alcohol, boolean artesanal, double graduacion){
         super(descripcion, nombre, marca, tipo, cantidad, precio, alcohol, artesanal, graduacion);
         this.color = color;
         this.ibu = ibu;
     }
 
-    public int getIbu() {
+    public double getIbu() {
         return ibu;
     }
